Make the hosted gate pass: rustfmt the tree, burn down stable-1.96 clippy
The ci gate has been red since the initial commit: the history-reset
squash landed a pre-rustfmt tree, so the very first step (cargo fmt
--check) failed and clippy/test never ran on hosted CI. rustfmt the
tree and clear the clippy -D warnings backlog the gate would have hit
next (stable 1.96 lints: manual_repeat_n, manual_is_multiple_of,
manual_div_ceil, manual_range_contains, map_or simplifications,
needless_range_loop, ptr_arg, doc_lazy_continuation, collapsible
match arm, suspicious_open_options).

All rewrites are behavior-preserving; the two production open sites
now pin .truncate(false) explicitly (the previous default), matching
the uring engine's precedent, and the catalog reload uses
unwrap_or_default() (NameCatalog is a BTreeMap alias, default ==
new). Structural lints on the careful-semantics surfaces
(too_many_arguments on the split path, type_complexity on the range
iterators, wrong_self_convention on GroupSerializer::from_slice) are
allowed rather than refactored. Local gate: fmt clean, clippy clean
on both crates, cargo test --locked all green.

+++ tests/ser_columnar.rs
@@ -264,7 +264,7 @@ fn column_cursor_reads_only_its_column() {
         .map(|i| {
             vec![
                 Value::Long(1000 + i as i64),
-                Value::Int(-(i as i32)),
+                Value::Int(-i),
                 Value::Short((i * 7) as i16),
             ]
         })
